+++ src/Backend/Modules/Pages/Engine/CacheBuilder.php
@@ -70,7 +70,7 @@ class CacheBuilder
     public function getKeys($language)
     {
         $item = $this->cache->getItem('keys_' . $language);
-        if ($item->exists()) {
+        if ($item->isHit()) {
             return $item->get();
         }
 
@@ -88,7 +88,7 @@ class CacheBuilder
     public function getNavigation($language)
     {
         $item = $this->cache->getItem('navigation_' . $language);
-        if ($item->exists()) {
+        if ($item->isHit()) {
             return $item->get();
         }
 
@@ -169,8 +169,8 @@ class CacheBuilder
             'page_id' => (int) $page['id'],
             'url' => $page['url'],
             'full_url' => $languageURL . $keys[$page['id']],
-            'title' => preg_replace('/\'/', '\\\\\'', $page['title']),
-            'navigation_title' => preg_replace('/\'/', '\\\\\'', $page['navigation_title']),
+            'title' => $page['title'],
+            'navigation_title' => $page['navigation_title'],
             'has_extra' => (bool) ($page['has_extra'] == 'Y'),
             'no_follow' => (bool) (isset($page['meta_data']['seo_follow']) && $page['meta_data']['seo_follow'] == 'nofollow'),
             'hidden' => (bool) ($page['hidden'] == 'Y'),