+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 --
 Bugfixes:
 * Pages: removed single quotes converter in CacheBuilder. This fixes page titles with single quotes.
+* Backend: Core fix when using extra-params in a datagrid sequence reorder ajax call.
 
 3.9.6 (2015-12-22)
 --