+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+3.9.7 (xxxx-xx-xx)
+--
+Bugfixes:
+* Faq: fixes faq-category sequence reordering not being saved.
+* Pages: removed single quotes converter in CacheBuilder. This fixes page titles with single quotes.
+
 3.9.6 (2015-12-22)
 --
 Improvements: